+++ plugin-dev/Source/Sentry/Private/Apple/AppleSentrySubsystem.cpp
@@ -253,10 +253,39 @@ void FAppleSentrySubsystem::InitWithSettings(const USentrySettings* settings, co
 	// Wait synchronously until sentry-cocoa initialization finished in main thread
 	dispatch_group_wait(sentryDispatchGroup, DISPATCH_TIME_FOREVER);
 	dispatch_release(sentryDispatchGroup);
+
+	if (IsEnabled())
+	{
+#ifdef USE_SENTRY_SESSION_REPLAY
+		if (settings->AttachSessionReplay)
+		{
+			SessionReplayId = FGuid::NewGuid().ToString(EGuidFormats::Digits).ToLower();
+
+			SessionReplay = MakeUnique<FSentrySessionReplayRecorder>();
+			if (SessionReplay->Initialize(settings, SessionReplayId, GetReplayPath()))
+			{
+				SetContext(TEXT("replay"), { { TEXT("replay_id"), FSentryVariant(SessionReplayId) } });
+			}
+			else
+			{
+				SessionReplay.Reset();
+				SessionReplayId.Reset();
+			}
+		}
+#endif
+	}
 }
 
 void FAppleSentrySubsystem::Close()
 {
+#ifdef USE_SENTRY_SESSION_REPLAY
+	if (SessionReplay)
+	{
+		SessionReplay->Shutdown();
+		SessionReplay.Reset();
+	}
+#endif
+
 	[SENTRY_APPLE_CLASS(SentryObjCSDK) flush:0];
 	[SENTRY_APPLE_CLASS(SentryObjCSDK) close];
 }
@@ -798,5 +827,13 @@ FString FAppleSentrySubsystem::GetLatestSessionReplay() const
 
 	return Replays[0];
 }
+
+#ifdef USE_SENTRY_SESSION_REPLAY
+FString FAppleSentrySubsystem::GetReplayPath() const
+{
+	const FString ReplayPath = FPaths::Combine(FPaths::ProjectSavedDir(), TEXT("SentryReplays"), FString::Printf(TEXT("replay-%s.mp4"), *SessionReplayId));
+	return FPaths::ConvertRelativePathToFull(ReplayPath);
+}
+#endif
 
 #endif // !USE_SENTRY_NATIVE